+++ app/src/main/java/com/example/compassapp/data/models/Coordinate.kt
@@ -1,3 +1,3 @@
 package com.example.compassapp.data.models
 
-class Coordinate(val latitude: Double, val longitude: Double)
+class Coordinate(val latitude: Float, val longitude: Float)
