+++ app/src/main/java/com/example/compassapp/ui/UpdateDirectionDialog.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
 import com.example.compassapp.R
-import com.example.compassapp.utils.getDoubleValue
+import com.example.compassapp.utils.getFloatValue
 import com.example.compassapp.utils.isEmpty
 import com.example.compassapp.utils.showErrorToastLong
 import com.example.compassapp.viewmodels.CompassViewModel
@@ -37,8 +37,8 @@ class UpdateDirectionDialog : DialogFragment() {
     private fun setListeners() {
         disposable = btShowDirection.clicks().subscribe {
             if (areEnteredValuesCorrect()) {
-                val latitude = etLatitude.getDoubleValue()
-                val longitude = etLongitude.getDoubleValue()
+                val latitude = etLatitude.getFloatValue()
+                val longitude = etLongitude.getFloatValue()
                 viewModel.saveChosenCoordinates(latitude, longitude)
                 dismiss()
             } else {
@@ -60,8 +60,8 @@ class UpdateDirectionDialog : DialogFragment() {
 
     private fun areEnteredValuesCorrect(): Boolean {
         if (etLatitude.isEmpty() || etLongitude.isEmpty()) return false
-        if (90 < etLatitude.getDoubleValue() || etLatitude.getDoubleValue() < -90) return false
-        if (180 < etLongitude.getDoubleValue() || etLongitude.getDoubleValue() < -180) return false
+        if (90 < etLatitude.getFloatValue() || etLatitude.getFloatValue() < -90) return false
+        if (180 < etLongitude.getFloatValue() || etLongitude.getFloatValue() < -180) return false
         return true
     }
